SSE conversion from UTF-16 to UTF-8
Handle surrogate pairs in SIMD code. Also added extensively
tests that cover all cases for SSE code.

+++ tests/helpers/random_int.cpp
@@ -14,10 +14,10 @@ namespace simdutf::tests::helpers {
       return distr(gen);
     }
 
-    RandomIntRanges::RandomIntRanges(std::initializer_list<std::pair<uint64_t, uint16_t>> ranges)
+    RandomIntRanges::RandomIntRanges(std::initializer_list<std::pair<uint64_t, uint64_t>> ranges)
       : RandomIntRanges(ranges, std::random_device{}()) {}
 
-    RandomIntRanges::RandomIntRanges(std::initializer_list<std::pair<uint64_t, uint16_t>> ranges_, uint64_t seed)
+    RandomIntRanges::RandomIntRanges(std::initializer_list<std::pair<uint64_t, uint64_t>> ranges_, uint64_t seed)
       : gen(seed) {
 
       if (ranges_.size() == 0)
